requirements.txt
cohere>=5.11.3
lancedb>=0.15.0
google-generativeai>=0.8.3
sentence-transformers>=3.2.1  # apparently needed for even cohere-reranker api
streamlit>=1.39.0